requirements.txt
# ============================================
# HelloDoc Backend - Requirements
# ============================================

# Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0

# Database
psycopg[binary]>=3.2,<4.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment & Configuration
python-decouple>=3.8,<4.0
dj-database-url>=2.2,<3.0

# Storage (Supabase S3-compatible)
django-storages>=1.14,<2.0
boto3>=1.35,<2.0

# Background Tasks
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0

# OpenAI (Whisper + GPT-4)
openai>=1.50,<2.0

# Export
reportlab>=4.2,<5.0

# Utilities
python-dateutil>=2.9,<3.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0

# Dev & Testing
ruff>=0.6,<1.0
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
